Fixing some warnings

+++ modules/tracking/samples/kcf.cpp
@@ -165,16 +165,16 @@ Rect2d BoxExtractor::extract(const std::string& windowName, Mat img, bool showCr
       // horizontal line
       line(
         params.image,
-        Point(params.box.x,params.box.y+(int)(params.box.height/2)),
-        Point(params.box.x+params.box.width,params.box.y+(int)(params.box.height/2)),
+        Point((int)params.box.x,(int)(params.box.y+params.box.height/2)),
+        Point((int)(params.box.x+params.box.width),(int)(params.box.y+params.box.height/2)),
         Scalar(255,0,0),2,1
       );
 
       // vertical line
       line(
         params.image,
-        Point(params.box.x+(int)(params.box.width/2),params.box.y),
-        Point(params.box.x+(int)(params.box.width/2),params.box.y+params.box.height),
+        Point((int)(params.box.x+params.box.width/2),(int)params.box.y),
+        Point((int)(params.box.x+params.box.width/2),(int)(params.box.y+params.box.height)),
         Scalar(255,0,0),2,1
       );
     }
